+++ src/main/java/ru/skillbox/social_network_post/service/impl/ReactionServiceImpl.java
@@ -67,7 +67,7 @@ public class ReactionServiceImpl implements ReactionService {
             postRepository.incrementReactionsCount(postId);
         }
 
-        int count = reactionRepository.countByPostIdAndReactionType(postId, requestReactionDto.reactionType());
+        long count = reactionRepository.countByPostIdAndReactionType(postId, requestReactionDto.reactionType());
 
         ReactionNotificationDto reactionNotificationDto = ReactionNotificationDto.builder()
                 .authorId(accountId)
